scripts/LivePopularTimes/pyproject.toml
[tool.poetry]
name = "livepopulartimes"
version = "0.1.0"
description = "get popular times for a restaurant"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
python-dotenv = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
